--- a/lib/segment/src/index/field_index/full_text_index/immutable_text_index.rs
+++ b/lib/segment/src/index/field_index/full_text_index/immutable_text_index.rs
@@ -85,16 +85,17 @@ impl ImmutableFullTextIndex {
     }
 
     pub fn remove_point(&mut self, id: PointOffsetType) -> OperationResult<()> {
-        match self.storage {
-            Storage::RocksDb(ref db_wrapper) => {
-                let db_doc_id = FullTextIndex::store_key(id);
-                db_wrapper.remove(db_doc_id)?;
-            }
-            Storage::Mmap(ref mut index) => {
-                index.remove_point(id)?;
+        if self.inverted_index.remove_document(id) {
+            match self.storage {
+                Storage::RocksDb(ref db_wrapper) => {
+                    let db_doc_id = FullTextIndex::store_key(id);
+                    db_wrapper.remove(db_doc_id)?;
+                }
+                Storage::Mmap(ref mut index) => {
+                    index.remove_point(id)?;
+                }
             }
         }
-
         Ok(())
     }